Update lc-other-settings.md
Customer contact was unable to find this documentation when searching for details about adding snippets in AEM guides and what permissions are necessary. 
They state: " I had looked for this in search / the TOC but didn't think to look at the "Configure other settings" result (generically named) or at anything under "Manage Learning Content" since we're not using any of the e-learning functionality at this time, just standard DITA / documentation features. This article could be better titled and filed"

I have added an update to the title and added a sub header for managing snippets.